+++ src/main/kotlin/one/pkg/modpublish/util/metadata/ModTomlParser.kt
@@ -18,48 +18,58 @@ package one.pkg.modpublish.util.metadata
 
 import com.intellij.openapi.diagnostic.Logger
 import one.pkg.modpublish.data.internal.LocalModInfo
+import one.pkg.modpublish.data.internal.SideType
+import one.pkg.modpublish.util.io.toml.KTomlArray
 import one.pkg.modpublish.util.io.toml.KTomlParser
 import java.io.InputStream
 import java.lang.AutoCloseable
 
 @Suppress("UNUSED")
 data class ModTomlParser(val parser: KTomlParser) : AutoCloseable {
-    fun get(): LocalModInfo? {
-        return runCatching {
-            val mods = parser.getAsTomlParser("mods")
-            if (mods == null || mods.isEmpty()) {
-                LOG.warn("No mods section found in TOML file")
-                return null
-            }
-
-            val displayName = mods.getAsString("displayName")
-            val version = mods.getAsString("version")
-
-            LocalModInfo(displayName, version, getMinecraftVersions(mods.getAsString("modId")))
-        }.onFailure { LOG.error("Failed to parse mod.toml", it) }.getOrNull()
-    }
-
-    fun getMinecraftVersions(modId: String): String {
-        val dependencies = parser.getAsTomlParser("dependencies")
-        if (dependencies != null) {
-            val arrays = dependencies.getAsTomlArray(modId)
-            if (arrays.isEmpty) return ""
-            for (array in arrays)
-                if (array.getAsString("modId") == "minecraft") return array.getAsString("versionRange")
+    fun get(): LocalModInfo? = runCatching {
+        val mods = parser.getAsTomlParser("mods")?.takeUnless { it.isEmpty() } ?: run {
+            LOG.warn("No mods section found in TOML file")
+            return null
         }
 
-        return ""
-    }
+        val deps = getArray(mods.getAsString("modId"))
+        val clientSide = if (parser.getAsBoolean("clientSideOnly")) SideType.CLIENT else SideType.BOTH
 
-    override fun close() {
-        // NONE
+        LocalModInfo(
+            name = mods.getAsString("displayName"),
+            version = mods.getAsString("version"),
+            versionRange = deps?.let(::getMinecraftVersions).orEmpty(),
+            sideType = if (clientSide == SideType.CLIENT) SideType.CLIENT else deps?.let(::getSideType) ?: SideType.BOTH
+        )
+    }.onFailure { LOG.error("Failed to parse mod.toml", it) }.getOrNull()
+
+    private fun getSideType(arrays: KTomlArray): SideType =
+        arrays.firstOrNull { it.getAsString("modId") in setOf("minecraft", "forge", "neoforge") }
+            ?.getAsString("side")
+            ?.let {
+                when (it) {
+                    "BOTH" -> SideType.BOTH
+                    "CLIENT" -> SideType.CLIENT
+                    else -> SideType.SERVER
+                }
+            } ?: SideType.BOTH
+
+    private fun getArray(modId: String): KTomlArray? =
+        parser.getAsTomlParser("dependencies")
+            ?.getAsTomlArray(modId)
+            ?.takeUnless { it.isEmpty }
+
+    private fun getMinecraftVersions(arrays: KTomlArray): String =
+        arrays.firstOrNull { it.getAsString("modId") == "minecraft" }
+            ?.getAsString("versionRange")
+            ?: ""
+
+    override fun close() { /* no-op */
     }
 
     companion object {
         private val LOG = Logger.getInstance(ModTomlParser::class.java)
 
-        fun InputStream.toModTomlParser(): ModTomlParser {
-            return ModTomlParser(KTomlParser.fromStream(this))
-        }
+        fun InputStream.toModTomlParser(): ModTomlParser = ModTomlParser(KTomlParser.fromStream(this))
     }
 }