+++ changelogs/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Reimplement Gitlab
 - Support sync README to Modrinth
 - Support LiteLoader and JavaAgent
+- Support automatic selection of Support Targets
 
 ### Fix
 - Fixed the parsing failure caused by multiple ways of writing the minecraft version range in fabric.mod.json